CA-104 #comment Improve Emitters, Introduce BufferChains
BufferChains are sinlgy linked lists of immuteable Buffers.
They replace the std::vector<std::string> implementation.

Emitters now hold a BinaryBuilder that appends data until
a) buffer full or
b) user calls flush
In Both cases a BinaryBuffer instance is appended to the target.
A target can either be a BufferChain or a network socket.

Iterators hide the fact that there is a BufferChain in the background by
providing a smooth HasNext() / Next() interface without hassle for the user.

This commit breaks some tests that I have commented out, but I gonna fix them
in the follow-up commits