CA-104 #comment amend hash table tests
CA-71 #comment I had to change the tets because the emitter is not a lambda any more. sorry that the stuff is not printed on the std:cout any more, if you need to see the values, you can use the iterator to access the stuff that got pushed throught the emitter.
CA-71 #comment you can now call 'Close' on emitters to signal that all data has been sent