Add network with topology creation from function
The Network class provides a collection of Groups which are structured in a
given topology. The function which creates the topology is given as a parameter
to the Network constructor, thereby making it possible to specify an own
topology, and just passing the creation function to the Network. The only
currently supported topology is the linked list. Other topologies might include
a fibonacci-tree or other data structures for efficient communication.